+++ examples/advanced/never-miss-a-beat.ts
@@ -1,5 +1,15 @@
 import { DFUSE_API_KEY, runMain, DFUSE_API_NETWORK } from "../config"
-import { createDfuseClient, InboundMessage, waitFor, Stream } from "@dfuse/client"
+import {
+  createDfuseClient,
+  InboundMessage,
+  waitFor,
+  Stream,
+  DfuseClient,
+  dynamicMessageDispatcher,
+  ProgressInboundMessage,
+  ActionTraceInboundMessage,
+  Action
+} from "@dfuse/client"
 
 /**
  * In this example, we will showcase how to implement bullet proof
@@ -14,7 +24,12 @@ import { createDfuseClient, InboundMessage, waitFor, Stream } from "@dfuse/clien
  * and how the marker is then used when the socket re-connects to
  * restart the stream at the exact location you need.
  *
- * @see
+ * In the example. we will implement an action persistence storer
+ * showing our to restart at the exact correct place a commit had
+ * occurred.
+ *
+ *
+ * @see https://docs.dfuse.io/#websocket-based-api-never-missing-a-beat
  */
 async function main() {
   const client = createDfuseClient({
@@ -22,45 +37,175 @@ async function main() {
     network: DFUSE_API_NETWORK,
     streamClientOptions: {
       socketOptions: {
+        onError(error) {
+          console.log("An error occurred", error)
+        },
         onReconnect() {
-          /**
-           * This listener is called by the socket right after a connection
-           * has be re-open with the server. This is not called on the initial
-           * connection.
-           *
-           * It's in this method that you should re-connect all your stream
-           * manually. In this sample, simply restart the stream.
-           */
-          console.log("Socket has been reconnected with remote server, restarting stream.")
-          stream.restart()
+          console.log()
+          console.log("<============= Stream as re-connected to socket correctly =============>")
+          console.log()
+
+          // Upon a re-connection, we need to clear previously accumulated actions
+          engine.flushPending()
         }
       }
     }
   })
 
-  const stream: Stream = await client.streamActionTraces(
-    {
-      account: "eosio.token",
-      action_name: "create"
-    },
-    onMessage
-  )
+  const engine = new Engine(client)
+  await engine.start()
 
-  console.log("Socket is now connected.")
-
-  await waitFor(38000)
-  await stream.close()
+  await waitFor(50000)
+  await engine.stop()
 }
 
-function onMessage(message: InboundMessage) {
-  if (message.type === "listening") {
-    console.log("Stream is now listening.")
-    return
+type KarmaTransfer = {
+  from: string
+  to: string
+  quantity: string
+  memo: string
+}
+
+class Engine {
+  private client: DfuseClient
+  private stream?: Stream
+
+  private pendingActions: Array<Action<KarmaTransfer>> = []
+  private lastCommittedBlockNum: number = 0
+
+  private committedActions: Array<Action<KarmaTransfer>> = []
+
+  constructor(client: DfuseClient) {
+    this.client = client
   }
 
-  if (message.type === "action_trace") {
-    console.log("Streaming transfer.")
+  public async start() {
+    const dispatcher = dynamicMessageDispatcher({
+      listening: this.onListening,
+      action_trace: this.onAction,
+      progress: this.onProgress
+    })
+
+    console.log("Engine starting")
+    this.stream = await this.client.streamActionTraces(
+      {
+        account: "therealkarma",
+        action_name: "transfer"
+      },
+      dispatcher,
+      {
+        // You can use the `with_progress` to be sure to commit
+        // actions at least each 10 blocks. This is useful if your stream
+        // is low traffic so you don't need to wait until next
+        // action to commit all changes.
+        with_progress: 10
+      }
+    )
+
+    console.log("Stream connected, ready to receive messages")
   }
+
+  public async stop() {
+    await this.ensureStream().close()
+
+    console.log("Committed actions")
+    this.committedActions.forEach((action) => {
+      const { from, to, quantity } = action.data
+      console.log(`- Commit transfer [${from} -> ${to} ${quantity}]`)
+    })
+  }
+
+  /**
+   * When the stream re-connects, we must flush all our current pending transactions
+   * as the stream re-starts at our last marked block, inclusive.
+   *
+   * Since we mark after commit, anything currently in pending was not committed,
+   * hence let's flush all pending actions, dfuse Stream will stream them back.
+   */
+  public async flushPending() {
+    console.log("Flushing pending action(s) due to refresh")
+    this.pendingActions = []
+  }
+
+  private onListening = () => {
+    console.log("Stream is now listening for action(s)")
+  }
+
+  private onProgress = (message: ProgressInboundMessage) => {
+    const { block_id, block_num } = message.data
+
+    /**
+     * Once a progress message is seen, it means we've seen all messages for
+     * block prior it, so le'ts commit until this point.
+     */
+    console.log()
+    console.log("Committing changes due to seeing a message from a progress message")
+    this.commit(block_id, block_num)
+  }
+
+  private onAction = (message: ActionTraceInboundMessage<KarmaTransfer>) => {
+    /**
+     * Once a message from a block ahead of last committed block is seen,
+     * commit all changes up to this point.
+     */
+    const { block_id, block_num } = message.data
+    if (block_num > this.lastCommittedBlockNum) {
+      console.log()
+      console.log(
+        "Comitting changes due to seeing a message from a block ahead of our last committed block"
+      )
+      this.commit(block_id, block_num)
+    }
+
+    const action = message.data.trace.act
+    const { from, to, quantity } = action.data
+
+    console.log(
+      `Pending transfer [${from} -> ${to} ${quantity}] @ ${printBlock(block_id, block_num)}`
+    )
+    this.pendingActions.push(message.data.trace.act)
+  }
+
+  private commit(blockId: string, blockNum: number) {
+    console.log(`Committing all actions up to block ${printBlock(blockId, blockNum)}`)
+
+    if (this.pendingActions.length > 0) {
+      // Here, in your production code, action would be saved in a database, as well as error handling
+      this.pendingActions.forEach((action) => this.committedActions.push(action))
+    }
+
+    console.log(`Bumping last committed block and clearing pending actions`)
+    this.pendingActions = []
+    this.lastCommittedBlockNum = blockNum
+
+    /**
+     * This is one of the most important call of the example. By marking the stream
+     * at the right block, upon restart, the stream will automatically starts back
+     * at this block ensuring to never miss a single action.
+     */
+    console.log(`Marking stream up to block ${printBlock(blockId, blockNum)}`)
+    this.ensureStream().mark({ atBlockNum: blockNum })
+
+    /**
+     * In a real-word production code, you would also need to persist the
+     * `this.lastCommittedBlockNum` value to ensure that upon a process
+     * restart, you start back from this exact value.
+     */
+
+    console.log("")
+  }
+
+  private ensureStream(): Stream {
+    if (this.stream) {
+      return this.stream
+    }
+
+    throw new Error("Stream should be set at this runtime execution point")
+  }
+}
+
+function printBlock(blockId: string, blockNum: number): string {
+  return `${blockId.slice(0, 8)}...${blockId.slice(-8)} (${blockNum})`
 }
 
 runMain(main)